requirements.txt
pillow
requests
termcolor